requirements.txt
langchain==0.0.267
openai==0.28.1
python-dotenv
tiktoken
sentence-transformers
faiss-cpu
pypdf